Update ASML_spec.mcrl2
Added bunch of variables to check for ocupation and in every step of the controller it does one thing and tells the physical device to also do that. Need to add IPU and more Robot pick, put down and the Other Operation modus. Also Comments everywhere right now to big a blob of text.